postgres_to_es/requirements.txt
python-dotenv==0.21.0
psycopg2-binary==2.9.5
pydantic==2.6.0
elasticsearch==8.6.2
asyncio
pydantic_settings